requirements.txt
Django==5.0.7
sentry-sdk==2.11.0
djangorestframework==3.15.2
django-cors-headers==4.4.0
